Add bench_executor
The bench_executor is responsible for looping through all required
parameter combinations, orchestrating the amount of sensors in the
benchmarking system and invoking the test_driver with the correct
parameters.
This involves repeating each run n times, and passing the ips to the
sensors to the test driver.

For now, invoking the test driver is not implemented, only the
configuration of the number of sensors in the system